refactor: move DiscussionHighlight to discussion.highlight package
