+++ chaekit-spring/src/main/java/qwerty/chaekit/domain/group/activity/discussion/Discussion.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.BatchSize;
 import qwerty.chaekit.domain.BaseEntity;
-import qwerty.chaekit.domain.discussionhighlight.DiscussionHighlight;
+import qwerty.chaekit.domain.group.activity.discussion.highlight.DiscussionHighlight;
 import qwerty.chaekit.domain.group.activity.Activity;
 import qwerty.chaekit.domain.group.activity.discussion.comment.DiscussionComment;
 import qwerty.chaekit.domain.highlight.entity.Highlight;
