+++ chaekit-spring/src/main/java/qwerty/chaekit/service/highlight/HighlightService.java
@@ -7,8 +7,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import qwerty.chaekit.domain.discussionhighlight.DiscussionHighlight;
-import qwerty.chaekit.domain.discussionhighlight.DiscussionHighlightRepository;
+import qwerty.chaekit.domain.group.activity.discussion.highlight.DiscussionHighlight;
+import qwerty.chaekit.domain.group.activity.discussion.highlight.DiscussionHighlightRepository;
 import qwerty.chaekit.domain.ebook.Ebook;
 import qwerty.chaekit.domain.group.activity.Activity;
 import qwerty.chaekit.domain.group.activity.discussion.Discussion;
